+++ modules/news/src/frontend/Module.php
@@ -15,6 +15,13 @@ class Module extends \luya\base\Module
     public $useAppViewPath = true;
 
     /**
+     * @var array The default order for the article overview in the index action for the news.
+     * 
+     * In order to read more about activeDataProvider defaultOrder: http://www.yiiframework.com/doc-2.0/yii-data-sort.html#$defaultOrder-detail
+     */
+    public $articleDefaultOrder = ['timestamp_create' => SORT_DESC];
+    
+    /**
      * @var array
      */
     public $urlRules = [