+++ modules/news/src/models/Article.php
@@ -76,17 +76,6 @@ class Article extends NgRestModel
             $this->timestamp_display_from = time();
         }
     }
-    
-    /**
-     * @inheritdoc
-     */
-    public function scenarios()
-    {
-        return [
-           'restcreate' => ['title', 'text', 'cat_id', 'image_id', 'image_list', 'timestamp_create', 'timestamp_display_from', 'timestamp_display_until', 'is_display_limit', 'file_list', 'teaser_text'],
-           'restupdate' => ['title', 'text', 'cat_id', 'image_id', 'image_list', 'timestamp_create', 'timestamp_display_from', 'timestamp_display_until', 'is_display_limit', 'file_list', 'teaser_text'],
-       ];
-    }
 
     /**
      * @inheritdoc
@@ -97,7 +86,7 @@ class Article extends NgRestModel
             [['title', 'text'], 'required'],
             [['title', 'text', 'image_list', 'file_list', 'teaser_text'], 'string'],
             [['cat_id', 'create_user_id', 'update_user_id', 'timestamp_create', 'timestamp_update', 'timestamp_display_from', 'timestamp_display_until', 'is_deleted', 'is_display_limit'], 'integer'],
-            [['image_id'], 'string', 'max' => 200],
+            [['image_id'], 'safe'],
         ];
     }